Created a RobotName class to store the names and avoid class variables :), also made it to a separate class because its not short.
